backend/requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.14,<3.15
django-cors-headers>=4.3,<4.4
djangorestframework-simplejwt>=5.3,<5.4
django-environ>=0.11,<0.12
django-filter>=23.5,<24.0
django-extensions>=3.2,<3.3

# Geospatial
psycopg2-binary>=2.9,<2.10
pgvector>=0.2,<0.3

# Async
celery>=5.3,<5.4
redis>=5.0,<5.1
django-celery-beat>=2.6,<2.7
django-celery-results>=2.5,<2.6

# AWS
boto3>=1.34,<1.35

# Scraping (existing pipeline)
playwright>=1.41,<1.42
beautifulsoup4>=4.12,<4.13
requests>=2.31,<2.32

# Data processing
pandas>=2.2,<2.3
openpyxl>=3.1,<3.2
PyMuPDF>=1.23,<1.24
langchain-text-splitters>=0.0.1,<1.0

# Maps / Geo
geojson>=3.1,<3.2
shapely>=2.0,<2.1

# Notifications
twilio>=8.11,<8.12
firebase-admin>=6.4,<6.5

# Document generation
reportlab>=4.1,<4.2

# Dev
ipython>=8.0
